+++ kinto_ldap/__init__.py
@@ -14,6 +14,10 @@ DEFAULT_SETTINGS = {
                                   'LDAPBasicAuthAuthenticationPolicy'),
     'ldap.cache_ttl_seconds': 30,
     'ldap.endpoint': 'ldap://ldap.db.scl3.mozilla.com',
+    'ldap.pool_size': 10,
+    'ldap.pool_retry_max': 3,
+    'ldap.pool_retry_delay': .1,
+    'ldap.pool_timeout': 30,
     'ldap.fqn': 'mail={mail},o=com,dc=mozilla',
 }
 
@@ -36,4 +40,11 @@ def includeme(config):
 
     # Register heartbeat to ping the LDAP server.
     config.registry.heartbeats['ldap'] = ldap_ping
-    config.registry.ldap_cm = ConnectionManager(settings['ldap.endpoint'])
+
+    # LDAP pool connection manager
+    conn_options = dict(uri=settings['ldap.endpoint'],
+                        size=int(settings['ldap.pool_size']),
+                        retry_max=int(settings['ldap.pool_retry_max']),
+                        retry_delay=float(settings['ldap.pool_retry_delay']),
+                        timeout=int(settings['ldap.pool_timeout']))
+    config.registry.ldap_cm = ConnectionManager(**conn_options)